\name{linreg}

\alias{linreg-package}
\docType{package}
\title{
\packageTitle{linreg}
}
\description{
\packageDescription{linreg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{linreg}
\packageIndices{linreg}

}
\author{
\packageAuthor{linreg}

Maintainer: \packageMaintainer{linreg}
}
\references{
\url{https://en.wikipedia.org/wiki/Linear_regression}
}

\keyword{ package }
\seealso{
\url{https://github.com/poceviciute/AdvRprogr_lab4}

}
\examples{

}
