% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC_script.R
\docType{class}
\name{linreg}
\alias{linreg}
\title{linreg}
\arguments{
\item{formula}{a formula, format y ~ x_1 + x_2 + ... + x_n.}

\item{data}{a data frame.}
}
\value{
Returns the result of the Linear Regression
}
\description{
Returns the result of the Linear Regression
}

\examples{
data(iris)
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$print()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$plot()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$resid()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$pred()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$coef()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$summary()
}
