% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC_script.R
\docType{class}
\name{linreg}
\alias{linreg}
\title{linreg}
\arguments{
\item{formula}{a formula, format y ~ x_1 + x_2 + ... + x_n.}

\item{data}{a data frame.}
}
\value{
Returns the result of the Linear Regression
}
\description{
Returns the result of the Linear Regression
}

